+++ docs/ui/components/Home/index.tsx
@@ -1,3 +1,0 @@
-export * from './Cells';
-export * from './HomeButton';
-export * from './HeaderDescription';