+++ packages/expo-image/ios/ImageSource.swift
@@ -21,6 +21,9 @@ struct ImageSource: Record {
   @Field
   var cacheKey: String?
 
+  @Field
+  var useAppleWebpCodec: Bool = true
+
   var pixelCount: Double {
     return width * height * scale * scale
   }