+++ packages/expo-modules-core/expo-module-gradle-plugin/src/main/kotlin/expo/modules/plugin/ProjectConfiguration.kt
@@ -7,6 +7,8 @@ import expo.modules.plugin.android.PublicationInfo
 import expo.modules.plugin.android.applyLinterOptions
 import expo.modules.plugin.android.applyPublishingVariant
 import expo.modules.plugin.android.applySDKVersions
+import expo.modules.plugin.android.createEmptyExpoPublishTask
+import expo.modules.plugin.android.createEmptyExpoPublishToMavenLocalTask
 import expo.modules.plugin.android.createExpoPublishTask
 import expo.modules.plugin.android.createExpoPublishToMavenLocalTask
 import expo.modules.plugin.android.createReleasePublication
@@ -38,7 +40,10 @@ internal fun Project.applyKotlin(kotlinVersion: String, kspVersion: String) {
 internal fun Project.applyDefaultDependencies() {
   val modulesCore = rootProject.project(":expo-modules-core")
   if (project != modulesCore) {
-    project.dependencies.add("implementation", project.project(":expo-modules-core"))
+    project.dependencies.add("compileOnly", modulesCore)
+
+    project.dependencies.add("testImplementation", modulesCore)
+    project.dependencies.add("androidTestImplementation", modulesCore)
   }
 }
 
@@ -62,6 +67,8 @@ internal fun Project.applyDefaultAndroidSdkVersions() {
  */
 internal fun Project.applyPublishing(expoModulesExtension: ExpoModuleExtension) {
   if (!expoModulesExtension.canBePublished) {
+    createEmptyExpoPublishTask()
+    createEmptyExpoPublishToMavenLocalTask()
     return
   }
 
@@ -79,21 +86,14 @@ internal fun Project.applyPublishing(expoModulesExtension: ExpoModuleExtension) 
 
     createExpoPublishToMavenLocalTask(publicationInfo)
 
-    val publicationToken = rootProject.findProperty("EXPO_GITHUB_PUBLISH_TOKEN") as? String
-    if (!publicationToken.isNullOrEmpty()) {
-      publishingExtension().repositories.maven { mavenRepo ->
-        mavenRepo.name = "GitHubPackages"
-        mavenRepo.url = URI("https://maven.pkg.github.com/expo/expo")
-
-        mavenRepo.credentials { pc ->
-          pc.username = "expo"
-          pc.password = publicationToken
-        }
-      }
-      createExpoPublishTask(publicationInfo)
-    } else {
-      createExpoPublishTask(IllegalStateException("EXPO_GITHUB_PUBLISH_TOKEN is not defined"))
+    val npmLocalRepositoryRelativePath = "local-maven-repo"
+    val npmLocalRepository = URI("file://${project.projectDir.parentFile}/${npmLocalRepositoryRelativePath}")
+    publishingExtension().repositories.mavenLocal { mavenRepo ->
+      mavenRepo.name = "NPMPackage"
+      mavenRepo.url = npmLocalRepository
     }
+
+    createExpoPublishTask(publicationInfo, npmLocalRepositoryRelativePath)
   }
 }