+++ packages/expo-notifications/android/src/main/java/expo/modules/notifications/tokens/PushTokenModule.kt
@@ -1,42 +1,42 @@
 package expo.modules.notifications.tokens
 
-import android.os.Bundle
 import com.google.firebase.messaging.FirebaseMessaging
-import expo.modules.core.interfaces.services.EventEmitter
 import expo.modules.kotlin.Promise
 import expo.modules.kotlin.modules.Module
 import expo.modules.kotlin.modules.ModuleDefinition
-import expo.modules.notifications.ModuleNotFoundException
-import expo.modules.notifications.tokens.interfaces.PushTokenListener
-import expo.modules.notifications.tokens.interfaces.PushTokenManager
+import expo.modules.notifications.service.delegates.FirebaseMessagingDelegate.Companion.addTokenListener
+import expo.modules.notifications.tokens.interfaces.FirebaseTokenListener
 
 private const val NEW_TOKEN_EVENT_NAME = "onDevicePushToken"
 private const val NEW_TOKEN_EVENT_TOKEN_KEY = "devicePushToken"
 private const val REGISTRATION_FAIL_CODE = "E_REGISTRATION_FAILED"
 private const val UNREGISTER_FOR_NOTIFICATIONS_FAIL_CODE = "E_UNREGISTER_FOR_NOTIFICATIONS_FAILED"
 
-class PushTokenModule : Module(), PushTokenListener {
-  private val tokenManager: PushTokenManager? get() = appContext.legacyModuleRegistry
-    .getSingletonModule("PushTokenManager", PushTokenManager::class.java)
-  private var eventEmitter: EventEmitter? = null
+class PushTokenModule : Module(), FirebaseTokenListener {
+
+  init {
+    addTokenListener(this@PushTokenModule)
+  }
+
+  /**
+   * Callback called when [FirebaseMessagingDelegate] gets notified of a new token.
+   * Emits a [NEW_TOKEN_EVENT_NAME] event.
+   *
+   * @param token New push token.
+   */
+  override fun onNewToken(token: String) {
+    sendEvent(
+      NEW_TOKEN_EVENT_NAME,
+      mapOf(
+        NEW_TOKEN_EVENT_TOKEN_KEY to token
+      )
+    )
+  }
 
   override fun definition() = ModuleDefinition {
     Name("ExpoPushTokenManager")
 
     Events("onDevicePushToken")
-
-    OnCreate {
-      eventEmitter = appContext.legacyModule()
-        ?: throw ModuleNotFoundException(EventEmitter::class)
-
-      // Register the module as a listener in PushTokenManager singleton module.
-      // Deregistration happens in onDestroy callback.
-      tokenManager?.addListener(this@PushTokenModule)
-    }
-
-    OnDestroy {
-      tokenManager?.removeListener(this@PushTokenModule)
-    }
 
     /**
      * Fetches Firebase push token and resolves the promise.
@@ -86,20 +86,6 @@ class PushTokenModule : Module(), PushTokenListener {
         e
       )
       null
-    }
-  }
-
-  /**
-   * Callback called when [PushTokenManager] gets notified of a new token.
-   * Emits a [NEW_TOKEN_EVENT_NAME] event.
-   *
-   * @param token New push token.
-   */
-  override fun onNewToken(token: String) {
-    eventEmitter?.let {
-      val eventBody = Bundle()
-      eventBody.putString(NEW_TOKEN_EVENT_TOKEN_KEY, token)
-      it.emit(NEW_TOKEN_EVENT_NAME, eventBody)
     }
   }
 }