+++ packages/expo-notifications/src/scheduleNotificationAsync.ts
@@ -35,7 +35,7 @@ import {
  *     body: 'Change sides!',
  *   },
  *   trigger: {
- *     type: SchedulableTriggerInputTypes.TIME_INTERVAL,
+ *     type: Notifications.SchedulableTriggerInputTypes.TIME_INTERVAL,
  *     seconds: 60,
  *   },
  * });
@@ -50,7 +50,7 @@ import {
  *     title: 'Remember to drink water!',
  *   },
  *   trigger: {
- *     type: SchedulableTriggerInputTypes.TIME_INTERVAL,
+ *     type: Notifications.SchedulableTriggerInputTypes.TIME_INTERVAL,
  *     seconds: 60 * 20,
  *     repeats: true,
  *   },
@@ -155,7 +155,7 @@ function parseCalendarTrigger(
     trigger.type === SchedulableTriggerInputTypes.CALENDAR
   ) {
     const { repeats, ...calendarTrigger } = trigger;
-    return { type: 'calendar', value: calendarTrigger, repeats };
+    return { ...calendarTrigger, repeats: !!repeats, type: 'calendar' };
   }
   return undefined;
 }