+++ packages/expo-router/src/global-state/routing.ts
@@ -142,7 +142,9 @@ export type LinkToOptions = {
   dangerouslySingular?: SingularOptions;
 };
 
-export function linkTo(this: RouterStore, href: string, options: LinkToOptions = {}) {
+export function linkTo(this: RouterStore, originalHref: string, options: LinkToOptions = {}) {
+  let href: string | undefined = originalHref;
+
   if (emitDomLinkEvent(href, options)) {
     return;
   }
@@ -177,6 +179,12 @@ export function linkTo(this: RouterStore, href: string, options: LinkToOptions =
   const rootState = navigationRef.getRootState();
 
   href = resolveHrefStringWithSegments(href, this.routeInfo, options);
+  href = this.applyRedirects(href);
+
+  // If the href is undefined, it means that the redirect has already been handled the navigation
+  if (!href) {
+    return;
+  }
 
   const state = this.linking.getStateFromPath!(href, this.linking.config);