+++ packages/@expo/metro-config/CHANGELOG.md
@@ -10,6 +10,7 @@
 
 ### 💡 Others
 
+- Update RSC tests.
 - Add backup stack trace ([#35913](https://github.com/expo/expo/pull/35913) by [@EvanBacon](https://github.com/EvanBacon))
 
 ## 0.20.0 — 2025-04-04